CachedSubscription to its own header file
